Refractor for mappers + som cleanup
