+++ backend/src/main/java/com/tecnocampus/LS2/protube_back/application/dto/mapper/VideoMapper.java
@@ -5,6 +5,7 @@ import com.tecnocampus.LS2.protube_back.domain.video.VideoId;
 import com.tecnocampus.LS2.protube_back.application.dto.request.CreateVideoRequest;
 import com.tecnocampus.LS2.protube_back.application.dto.response.VideoResponse;
 import com.tecnocampus.LS2.protube_back.persistence.jpa.video.VideoProjection;
+import com.tecnocampus.LS2.protube_back.persistence.jpa.video.VideoEntity;
 import org.springframework.stereotype.Component;
 
 import java.time.Instant;
@@ -39,7 +40,7 @@ public class VideoMapper {
         );
     }
 
-    public Video toDomain(CreateVideoRequest request) {
+    public static Video toDomain(CreateVideoRequest request) {
         if (request == null) return null;
 
         Instant now = Instant.now();
@@ -65,19 +66,6 @@ public class VideoMapper {
                 now,
                 now
         );
-    }
-
-    private static String formatDuration(Integer seconds) {
-        if (seconds == null || seconds <= 0) return "00:00";
-
-        int minutes = seconds / 60;
-        int remainingSeconds = seconds % 60;
-        return String.format("%02d:%02d", minutes, remainingSeconds);
-    }
-
-    private static String formatFileSize(Long bytes) {
-        if (bytes == null || bytes <= 0) return null;
-        return String.format("%.1f MB", bytes / 1024.0 / 1024.0);
     }
 
     private static String buildMediaUrl(String path) {
@@ -114,4 +102,60 @@ public class VideoMapper {
     }
 
 
+
+    public static Video toDomain(VideoEntity entity) {
+        if (entity == null) {
+            return null;
+        }
+
+        return new Video(
+                new VideoId(entity.getId()),
+                entity.getJsonId(),
+                entity.getWidth(),
+                entity.getHeight(),
+                entity.getDurationSeconds(),
+                entity.getTitle(),
+                entity.getUser(),
+                entity.getTimestamp(),
+                entity.getDescription(),
+                entity.getCategories(),
+                entity.getTags(),
+                entity.getViewCount(),
+                entity.getLikeCount(),
+                entity.getChannel(),
+                CommentMapper.toDomainList(entity.getComments()),
+                entity.getMediaPath(),
+                entity.getThumbnailPath(),
+                entity.getCreatedAt(),
+                entity.getUpdatedAt()
+        );
+    }
+
+    public static VideoEntity toEntity(Video video) {
+        if (video == null) {
+            return null;
+        }
+
+        return new VideoEntity(
+                video.getId().value().toString(),
+                video.getJsonId(),
+                video.getWidth(),
+                video.getHeight(),
+                video.getDurationSeconds(),
+                video.getTitle(),
+                video.getUser(),
+                video.getTimestamp(),
+                video.getDescription(),
+                video.getCategories(),
+                video.getTags(),
+                video.getViewCount(),
+                video.getLikeCount(),
+                video.getChannel(),
+                CommentMapper.toEntityList(video.getComments()),
+                video.getMediaPath(),
+                video.getThumbnailPath(),
+                video.getCreatedAt(),
+                video.getUpdatedAt()
+        );
+    }
 }
